+++ nexusvault.format.tex/src/main/java/nexusvault/format/tex/struct/StructTextureFileHeader.java
@@ -16,7 +16,7 @@ import nexusvault.shared.exception.SignatureMismatchException;
 public final class StructTextureFileHeader {
 
 	public static final String STR_SIGNATURE = "GFX";
-	public static final int SIGNATURE = ('G' << 16) | ('F' << 8) | 'X';
+	public static final int SIGNATURE = 'G' << 16 | 'F' << 8 | 'X';
 
 	public static final int SIZE_IN_BYTES = StructUtil.sizeOf(StructTextureFileHeader.class); // 0x70;
 
@@ -80,7 +80,6 @@ public final class StructTextureFileHeader {
 	@StructField(BIT_32)
 	public int imageSizesCount; // 0x34
 
-	// public final int lengthOfCompressedImageSize;
 	@Order(13)
 	@StructField(value = BIT_32, length = 13)
 	public int[] imageSizes; // 0x38
@@ -92,6 +91,22 @@ public final class StructTextureFileHeader {
 	public StructTextureFileHeader() {
 	}
 
+	public StructTextureFileHeader(boolean initialized) {
+		if (!initialized) {
+			return;
+		}
+
+		this.signature = StructTextureFileHeader.SIGNATURE;
+		this.version = 1;
+		this.layerInfos = new StructLayerInfo[4];
+		for (int i = 0; i < this.layerInfos.length; ++i) {
+			this.layerInfos[i] = new StructLayerInfo();
+			this.layerInfos[i].hasReplacement(false);
+			this.layerInfos[i].setQuality(100);
+		}
+		this.imageSizes = new int[13];
+	}
+
 	public StructTextureFileHeader(BinaryReader reader) {
 		if (reader == null) {
 			throw new IllegalArgumentException("reader must not be null");
@@ -100,41 +115,41 @@ public final class StructTextureFileHeader {
 		final long headerStart = reader.getPosition();
 		final long headerEnd = headerStart + SIZE_IN_BYTES;
 
-		signature = reader.readInt32();
-		if (signature != StructTextureFileHeader.SIGNATURE) {
-			throw new SignatureMismatchException("Texture Header", SIGNATURE, signature);
+		this.signature = reader.readInt32();
+		if (this.signature != StructTextureFileHeader.SIGNATURE) {
+			throw new SignatureMismatchException("Texture Header", SIGNATURE, this.signature);
 		}
 
-		version = reader.readInt32();
-		width = reader.readInt32();
-		height = reader.readInt32();
-		depth = reader.readInt32();
-		sides = reader.readInt32();
-		mipMaps = reader.readInt32();
-		format = reader.readInt32();
+		this.version = reader.readInt32();
+		this.width = reader.readInt32();
+		this.height = reader.readInt32();
+		this.depth = reader.readInt32();
+		this.sides = reader.readInt32();
+		this.mipMaps = reader.readInt32();
+		this.format = reader.readInt32();
 
-		isCompressed = reader.readInt32() != 0;
-		compressionFormat = reader.readInt32();
+		this.isCompressed = reader.readInt32() != 0;
+		this.compressionFormat = reader.readInt32();
 
-		layerInfos = new StructLayerInfo[4];
-		for (int i = 0; i < layerInfos.length; ++i) {
-			layerInfos[i] = new StructLayerInfo(reader.readInt8(), reader.readInt8(), reader.readInt8());
-			if ((layerInfos[i].getQuality() < 0) || (100 < layerInfos[i].getQuality())) {
-				throw new IntegerOverflowException(layerInfos[i].toString());
+		this.layerInfos = new StructLayerInfo[4];
+		for (int i = 0; i < this.layerInfos.length; ++i) {
+			this.layerInfos[i] = new StructLayerInfo(reader.readInt8(), reader.readInt8(), reader.readInt8());
+			if (this.layerInfos[i].getQuality() < 0 || 100 < this.layerInfos[i].getQuality()) {
+				throw new IntegerOverflowException(this.layerInfos[i].toString());
 			}
 		}
 
-		imageSizesCount = reader.readInt32();
-		if (imageSizesCount > 12) {
+		this.imageSizesCount = reader.readInt32();
+		if (this.imageSizesCount > 12) {
 			throw new IllegalStateException("Number of compressed mips out of bound.");
 		}
 
-		imageSizes = new int[13];
-		for (int i = 0; i < imageSizes.length; ++i) {
-			imageSizes[i] = reader.readInt32();
+		this.imageSizes = new int[13];
+		for (int i = 0; i < this.imageSizes.length; ++i) {
+			this.imageSizes[i] = reader.readInt32();
 		}
 
-		unk_06C = reader.readInt32();
+		this.unk_06C = reader.readInt32();
 
 		if (reader.getPosition() != headerEnd) {
 			throw new IllegalStateException("Expected number of bytes " + SIZE_IN_BYTES + " read bytes: " + (reader.getPosition() - headerStart));
@@ -142,40 +157,40 @@ public final class StructTextureFileHeader {
 	}
 
 	public StructLayerInfo getLayer(int idx) {
-		return layerInfos[idx];
+		return this.layerInfos[idx];
 	}
 
 	@Override
 	public String toString() {
 		final StringBuilder builder = new StringBuilder();
 		builder.append("StructTextureFileHeader [signature=");
-		builder.append(signature);
+		builder.append(this.signature);
 		builder.append(", version=");
-		builder.append(version);
+		builder.append(this.version);
 		builder.append(", width=");
-		builder.append(width);
+		builder.append(this.width);
 		builder.append(", height=");
-		builder.append(height);
+		builder.append(this.height);
 		builder.append(", depth=");
-		builder.append(depth);
+		builder.append(this.depth);
 		builder.append(", sides=");
-		builder.append(sides);
+		builder.append(this.sides);
 		builder.append(", mipMaps=");
-		builder.append(mipMaps);
+		builder.append(this.mipMaps);
 		builder.append(", format=");
-		builder.append(format);
+		builder.append(this.format);
 		builder.append(", isCompressed=");
-		builder.append(isCompressed);
+		builder.append(this.isCompressed);
 		builder.append(", compressionFormat=");
-		builder.append(compressionFormat);
+		builder.append(this.compressionFormat);
 		builder.append(", layerInfos=");
-		builder.append(Arrays.toString(layerInfos));
+		builder.append(Arrays.toString(this.layerInfos));
 		builder.append(", imageSizesCount=");
-		builder.append(imageSizesCount);
+		builder.append(this.imageSizesCount);
 		builder.append(", imageSizes=");
-		builder.append(Arrays.toString(imageSizes));
+		builder.append(Arrays.toString(this.imageSizes));
 		builder.append(", unk_06C=");
-		builder.append(unk_06C);
+		builder.append(this.unk_06C);
 		builder.append("]");
 		return builder.toString();
 	}
@@ -184,10 +199,10 @@ public final class StructTextureFileHeader {
 	public int hashCode() {
 		final int prime = 31;
 		int result = 1;
-		result = (prime * result) + Arrays.hashCode(imageSizes);
-		result = (prime * result) + Arrays.hashCode(layerInfos);
-		result = (prime * result)
-				+ Objects.hash(compressionFormat, depth, format, height, imageSizesCount, isCompressed, mipMaps, sides, signature, unk_06C, version, width);
+		result = prime * result + Arrays.hashCode(this.imageSizes);
+		result = prime * result + Arrays.hashCode(this.layerInfos);
+		result = prime * result + Objects.hash(this.compressionFormat, this.depth, this.format, this.height, this.imageSizesCount, this.isCompressed,
+				this.mipMaps, this.sides, this.signature, this.unk_06C, this.version, this.width);
 		return result;
 	}
 
@@ -203,10 +218,10 @@ public final class StructTextureFileHeader {
 			return false;
 		}
 		final StructTextureFileHeader other = (StructTextureFileHeader) obj;
-		return (compressionFormat == other.compressionFormat) && (depth == other.depth) && (format == other.format) && (height == other.height)
-				&& Arrays.equals(imageSizes, other.imageSizes) && (imageSizesCount == other.imageSizesCount) && (isCompressed == other.isCompressed)
-				&& Arrays.equals(layerInfos, other.layerInfos) && (mipMaps == other.mipMaps) && (sides == other.sides) && (signature == other.signature)
-				&& (unk_06C == other.unk_06C) && (version == other.version) && (width == other.width);
+		return this.compressionFormat == other.compressionFormat && this.depth == other.depth && this.format == other.format && this.height == other.height
+				&& Arrays.equals(this.imageSizes, other.imageSizes) && this.imageSizesCount == other.imageSizesCount && this.isCompressed == other.isCompressed
+				&& Arrays.equals(this.layerInfos, other.layerInfos) && this.mipMaps == other.mipMaps && this.sides == other.sides
+				&& this.signature == other.signature && this.unk_06C == other.unk_06C && this.version == other.version && this.width == other.width;
 	}
 
 }