+++ nexusvault.format.tex/src/main/java/nexusvault/format/tex/unc/ARGB8888TextureImageReader.java
@@ -3,7 +3,7 @@ package nexusvault.format.tex.unc;
 public final class ARGB8888TextureImageReader extends UncompressedTextureImageReader {
 
 	public ARGB8888TextureImageReader() {
-		super(new ARGB8888Decoder(), new UncompressedImageMetaCalculator(4));
+		super(new ARGB8888ImageDecoder(), new UncompressedImageMetaCalculator(4));
 	}
 
 }