+++ nexusvault.format.tex/src/main/java/nexusvault/format/tex/unc/ARGB8888TextureImageWriter.java
@@ -1,9 +1,0 @@
-package nexusvault.format.tex.unc;
-
-public final class ARGB8888TextureImageWriter extends UncompressedTextureImageWriter {
-
-	public ARGB8888TextureImageWriter() {
-		super(new ARGB8888Encoder());
-	}
-
-}