+++ nexusvault.format.tex/src/main/java/nexusvault/format/tex/unc/Gray8TextureImageReader.java
@@ -3,7 +3,7 @@ package nexusvault.format.tex.unc;
 public final class Gray8TextureImageReader extends UncompressedTextureImageReader {
 
 	public Gray8TextureImageReader() {
-		super(new Gray8Decoder(), new UncompressedImageMetaCalculator(1));
+		super(new Gray8ImageDecoder(), new UncompressedImageMetaCalculator(1));
 	}
 
 }