+++ nexusvault.format.tex/src/main/java/nexusvault/format/tex/unc/Gray8TextureImageWriter.java
@@ -1,9 +1,0 @@
-package nexusvault.format.tex.unc;
-
-public final class Gray8TextureImageWriter extends UncompressedTextureImageWriter {
-
-	public Gray8TextureImageWriter() {
-		super(new Gray8Encoder());
-	}
-
-}