+++ nexusvault.format.tex/src/main/java/nexusvault/format/tex/unc/RGB565TextureImageReader.java
@@ -3,7 +3,7 @@ package nexusvault.format.tex.unc;
 public final class RGB565TextureImageReader extends UncompressedTextureImageReader {
 
 	public RGB565TextureImageReader() {
-		super(new RGB565Decoder(), new UncompressedImageMetaCalculator(2));
+		super(new RGB565ImageDecoder(), new UncompressedImageMetaCalculator(2));
 	}
 
 }