+++ nexusvault.format.tex/src/main/java/nexusvault/format/tex/unc/RGB565TextureImageWriter.java
@@ -1,9 +1,0 @@
-package nexusvault.format.tex.unc;
-
-public final class RGB565TextureImageWriter extends UncompressedTextureImageWriter {
-
-	public RGB565TextureImageWriter() {
-		super(new RGB565Encoder());
-	}
-
-}